trigger prerequisites from related make targets
While attempting to figure out how to add an exercise, I noticed that
some of the build targets defined in the Makefile were not automatic
(e.g. the generator build -> `generator/bin/generator`), so I added a
few prerequisites and their corresponding targets, along with a helper
target or two.

None of this is strictly necessary, of course. Please feel free to
close this PR with impunity if you see fit.